extract _plot_spectrum to FM_utils
